Create ViewModel - Adding Cart.cs and CartLine
Creating the viewmodels to support the UI.
The Cartclass uses the CartLineclass, defined in the same file, to represent a product and the quantity selected by the customer. I defined methods to add an item to the cart, remove a previously added item from the cart, calculate the total cost of the items in the cart, and reset the cart by removing all items. I also provided a property Lines that gives access to the contents of the cart using aList<CartLine>.